+++ src/Geometry.cpp
@@ -338,7 +338,7 @@ void GeometryGenerator::Subdivide(Mesh& mesh)
 	// *-----*-----*
 	// v0    m2     v2
 
-	UINT numTris = inputCopy.Indices.size()/3;
+	UINT numTris = (UINT)inputCopy.Indices.size()/3;
 	for(UINT i = 0; i < numTris; ++i)
 	{
 		Vertex v0 = inputCopy.Vertices[ inputCopy.Indices[i*3+0] ];
