+++ src/SphereRenderComponent.cpp
@@ -15,13 +15,13 @@ SphereRenderComponent::SphereRenderComponent(ID3D11VertexShader* pVertShader, ID
 	Mesh sphereMesh;
 	GeometryGenerator::CreateSphere(radius, sliceCount, stackCount, sphereMesh);
 
-	mIndexCount = sphereMesh.Indices.size();
+	mIndexCount = (UINT)sphereMesh.Indices.size();
 
 	D3DRenderer* renderer = Scene::getScene()->getRenderer();
 
 	D3D11_BUFFER_DESC bd;
 	bd.Usage = D3D11_USAGE_IMMUTABLE;
-	bd.ByteWidth = sizeof(Vertex) * sphereMesh.Vertices.size();
+	bd.ByteWidth = sizeof(Vertex) * (UINT)sphereMesh.Vertices.size();
 	bd.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 	bd.CPUAccessFlags = 0;
 	bd.MiscFlags = 0;
@@ -34,7 +34,7 @@ SphereRenderComponent::SphereRenderComponent(ID3D11VertexShader* pVertShader, ID
 
 	D3D11_BUFFER_DESC ibd;
 	ibd.Usage = D3D11_USAGE_IMMUTABLE;
-	ibd.ByteWidth = sizeof(UINT) * sphereMesh.Indices.size();
+	ibd.ByteWidth = sizeof(UINT) * (UINT)sphereMesh.Indices.size();
 	ibd.BindFlags = D3D11_BIND_INDEX_BUFFER;
 	ibd.CPUAccessFlags = 0;
 	ibd.MiscFlags = 0;
